Split out qMemoryRegion tests
